pyproject.toml
[tool.poetry]
name = "golfzon-ocr"
version = "0.1.0"
description = "A Python application that extracts player data from Golfzon scorecard screenshots using OCR and calculates net scores"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.28.0"
pytesseract = "^0.3.10"
Pillow = "^10.0.0"
opencv-python = "^4.8.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
